backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
supabase==2.3.5
beautifulsoup4==4.12.3
requests==2.31.0
playwright==1.42.0
APScheduler==3.10.4
pydantic==2.6.3
python-multipart==0.0.9
resend==0.7.2 